Moving from TravisCI to GH Actions for Python3.10 (#126)
For whatever reason, the TravisCI team doesn't have a working Python 3.10 build yet. At least, no one in exact parallel to their other Python builds. For this reason (and to reduce the number of online accounts I need to support this project) I am moving from TravisCI to the GitHub Acions workflows for mazelib. (There are other reasons I like this move.)

As such, I can now easily support Python 3.10. To do so, the required versions of NumPy and Cython had to be opened up, as those projects move to support Python3.10.

This also comes with a version bump.