Attempting to make the code coverage more predictable.
Currently, the code coverage randomly changes all the time,
because random number generators are heavily used to create mazes
and find solutions. But I want the code coverage to be more
predictable, so developers have meaningful feed back if their
PR reduces code coverage by a fraction of a percent.