Adding comments to MazeGenAlgo
Helping clarify the difference between lowercase and uppercase class attributes:

The difference between the lowercase and uppercase dimensions are:

* lowercase dimensions - does not include walls
* uppercase dimensions - includes walls

The reason this distinction exists is because walls can be as thick has hallways, or they can be paper thin. The difference is just one of how we DISPLAY the maze. For examples of how to display mazes with thicker or thinner walls, see the docs.